split large import records test file
